move SwaggerSchema to DiscoveryClient
